requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
chromadb>=0.4.18
sentence-transformers>=2.2.2
torch>=2.0.0
pypdf2>=3.0.1
python-docx>=1.1.0
python-dotenv>=1.0.0
pandas>=2.1.3
numpy>=1.26.2

# OCR support for scanned/image PDFs
pytesseract>=0.3.10
pdf2image>=1.17.0
Pillow>=10.0.0
# Optional: For API-based embeddings fallback
# openai>=1.0.0  # Uncomment if using OpenAI embeddings
# google-generativeai>=0.3.0  # Uncomment if using Gemini embeddings (note: Gemini doesn't have direct embeddings API)


